Refactor complete, tests passing
