+++ src/core/blocks/CMakeLists.txt
@@ -1,5 +1,5 @@
 message(STATUS "In Core Blocks: ${CMAKE_CURRENT_SOURCE_DIR}")
-add_library(core_block STATIC base_block.c)
+add_library(core_block STATIC base_block.c ser_block.c)
 target_include_directories(core_block PUBLIC ${INCLUDE_ALL})
 target_link_libraries(core_block
   core_tx)
